Add ConfigDirectoryContext framework

+++ amanuensis/config/loader.py
@@ -46,7 +46,9 @@ class ReadOnlyOrderedDict(OrderedDict):
 			raise AttributeError(key)
 		return self[key]
 
+
 class open_lock():
+	"""A context manager that opens a file with the specified file lock"""
 	def __init__(self, path, mode, lock_type):
 		self.fd = open(path, mode, encoding='utf8')
 		fcntl.lockf(self.fd, lock_type)
@@ -58,15 +60,25 @@ class open_lock():
 		fcntl.lockf(self.fd, fcntl.LOCK_UN)
 		self.fd.close()
 
+
 class open_sh(open_lock):
+	"""A context manager that opens a file with a shared lock"""
 	def __init__(self, path, mode):
 		super().__init__(path, mode, fcntl.LOCK_SH)
 
+
 class open_ex(open_lock):
+	"""A context manager that opens a file with an exclusive lock"""
 	def __init__(self, path, mode):
 		super().__init__(path, mode, fcntl.LOCK_EX)
 
+
 class json_ro(open_sh):
+	"""
+	A context manager that opens a file in a shared, read-only mode.
+	The contents of the file are read as JSON and returned as a read-
+	only OrderedDict.
+	"""
 	def __init__(self, path):
 		super().__init__(path, 'r')
 		self.config = None
@@ -75,17 +87,33 @@ class json_ro(open_sh):
 		self.config = json.load(self.fd, object_pairs_hook=ReadOnlyOrderedDict)
 		return self.config
 
+
 class json_rw(open_ex):
-	def __init__(self, path):
-		super().__init__(path, 'r+')
+	"""
+	A context manager that opens a file with an exclusive lock. The
+	file mode defaults to r+, which requires that the file exist. The
+	file mode can be set to w+ to create a new file by setting the new
+	kwarg in the ctor. The contents of the file are read as JSON and
+	returned in an AttrOrderedDict. Any changes to the context dict
+	will be written out to the file when the context manager exits.
+	"""
+	def __init__(self, path, new=False):
+		mode = 'w+' if new else 'r+'
+		super().__init__(path, mode)
 		self.config = None
+		self.new = new
 
 	def __enter__(self):
-		self.config = json.load(self.fd, object_pairs_hook=AttrOrderedDict)
+		if not self.new:
+			self.config = json.load(self.fd, object_pairs_hook=AttrOrderedDict)
+		else:
+			self.config = AttrOrderedDict()
 		return self.config
 
 	def __exit__(self, exc_type, exc_value, traceback):
-		self.fd.seek(0)
-		json.dump(self.config, self.fd, allow_nan=False, indent='\t')
-		self.fd.truncate()
+		# Only write the enw value out if there wasn't an exception
+		if not exc_type:
+			self.fd.seek(0)
+			json.dump(self.config, self.fd, allow_nan=False, indent='\t')
+			self.fd.truncate()
 		super().__exit__(exc_type, exc_value, traceback)
